chore: move blockchain module into event-svc crate
it will require a db connection so this avoids circular deps, plus it doesn't need to be in the validation crate
