+++ validation/Cargo.toml
@@ -10,6 +10,7 @@ publish = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+alloy.workspace = true
 anyhow.workspace = true
 async-trait.workspace = true
 base64.workspace = true
@@ -19,12 +20,10 @@ chrono.workspace = true
 ed25519-dalek.workspace = true
 ipld-core.workspace = true
 k256.workspace = true
-lru.workspace = true
 multibase.workspace = true
 multihash-codetable.workspace = true
 once_cell.workspace = true
 p256.workspace = true
-reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_ipld_dagcbor.workspace = true
@@ -33,8 +32,6 @@ ssi.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 unsigned-varint.workspace = true
-# will probably need contract and signer-local features for self anchoring
-alloy = { version = "0.4", features = ["k256", "provider-http", "rpc-types"] }
 
 [dev-dependencies]
 ceramic-car.workspace = true
