+++ validation/src/lib.rs
@@ -2,13 +2,11 @@
 //! as well as interacting with blockchains for time event validation.
 #![warn(missing_docs)]
 
-mod blockchain;
 mod signature;
 mod siwx_message;
 #[cfg(test)]
 mod test;
 mod verifier;
 
-pub use blockchain::eth_rpc;
 pub use ceramic_event::unvalidated::signed::cacao;
 pub use verifier::{cacao_verifier, event_verifier, AtTime, VerifyCacaoOpts, VerifyJwsOpts};
